Various moves and edits to fix cats unit tests
There are now redundant resources but we will clean them up next time
